Помогите

+++ client-app/Hbank/app/src/main/java/com/example/h_bank/data/Account.kt
@@ -5,7 +5,7 @@ import java.time.LocalDateTime
 data class Account(
     val id: String,
     val accountNumber: String,
-    val balance: Float,
+    val balance: Double,
     val userId: String,
     val isDeleted: Boolean,
     val createDateTime: LocalDateTime
